Move getFileMetas to separate file under util directory
