+++ server/script/migrateAssetsLocation.js
@@ -8,6 +8,7 @@ const {
   sequelize
 } = require('../shared/database');
 const get = require('lodash/get');
+const getFileMetas = require('../shared/util/getFileMetas');
 const Listr = require('listr');
 const path = require('path');
 const Promise = require('bluebird');
@@ -44,13 +45,13 @@ migrate()
 
 async function migrate() {
   const transaction = await sequelize.transaction();
-  const metaBySchemaType = getFileMetas(SCHEMAS);
-  const tasks = await getTasks(metaBySchemaType, transaction);
+  const tasks = await getTasks(transaction);
   return tasks.run().then(() => transaction.commit());
 }
 
-async function getTasks(metaBySchemaType, transaction) {
+async function getTasks(transaction) {
   const repositories = await Repository.findAll({ transaction });
+  const metaBySchemaType = getFileMetas(SCHEMAS);
   const tasks = repositories.map(repository => ({
     title: `Migrate repository "${repository.name}"`,
     task: () => {
@@ -251,39 +252,6 @@ class RepositoryMigration {
     }, {});
     return { ...metaInputs, ...newMeta };
   }
-}
-
-function getFileMetas(schemas) {
-  return schemas.reduce((acc, { id, meta, structure, elementMeta }) => {
-    return {
-      ...acc,
-      [id]: {
-        repository: getFileMetaKeys(meta),
-        activity: getMetaByActivityType(structure),
-        element: getMetaByElementType(elementMeta)
-      }
-    };
-  }, {});
-}
-
-function getMetaByActivityType(structure = []) {
-  return structure.reduce((acc, { type, meta }) => {
-    const fileMetaKeys = getFileMetaKeys(meta);
-    if (!fileMetaKeys.length) return acc;
-    return { ...acc, [type]: fileMetaKeys };
-  }, {});
-}
-
-function getMetaByElementType(elementMeta = []) {
-  return elementMeta.reduce((acc, { type, inputs }) => {
-    const fileMetaKeys = getFileMetaKeys(inputs);
-    if (!fileMetaKeys.length) return acc;
-    return { ...acc, [type]: fileMetaKeys };
-  }, {});
-}
-
-function getFileMetaKeys(meta = []) {
-  return meta.filter(it => it.type === 'FILE').map(it => it.key);
 }
 
 function resolveNewURL(assetUrl, targetDir) {
